Wrote a new custom JSON tokenizer replacement.
The old tokenizer was an attempt to reuse the scanner
included in the stdlib for golang itself.  As the goals
are different between JSONSM and the stdlib parser,
there were many opportunities for improvement.  This
rewritten version has a 2.5x tokenization speed
improvement which equates to about 2x filtering speed
increase in the current standard JSONSM benchmark.