Fixes #23: Try the alloc checker multiple times to determine success.
This is required due to the fact that Go can arbitrarily perform
allocations at any time, which causes the test to fail occasionally.
This commit causes it to run up to 20 times trying to get a 0-alloc
run, if any run has no allocations, it can be assumed that any other
allocations that were seen were arbitrary and unrelated to the binTree.