Fixes #37: Added new loop node type to the bintree.
In order to guarantee that the stall indexes used for loops,
and thus to guarantee that bucket 0 in the binary tree will
never be set to true during a loop, a new bintree node type
called loop has been added.  This element is inserted whenever
a loop is encountered.  This also aids with readability of
the binary tree.